Remove witness retrying logic on missing block
Sometimes when we would like to process a ChunkStateWitness the
block required to process it isn't available yet, we have to wait
for it. https://github.com/near/nearcore/pull/10535 implemented a
hacky way to do it by retrying the processing every 500ms until the
required block arrives. This PR will implement a proper solution, so
let's remove the hacky workaround.