Add verify_chunk_state_witness_signature_in_epoch
The current function to verify state witness signatures
takes the epoch_id from the previous block. For orphaned
state witnesses the previous block isn't available, so
it's impossible to use this function. Let's add another function
that takes the epoch_id as an argument instead of fetching it
from a block. It will allow us to verify signatures of orphaned
chunk state witnesses.