Add epoch_id_from_height_around_tip
Add a function which allows to determine whether a block height
belongs to one of the epochs around the chain tip. It achieves
this by comparing the height with height ranges covered by each
epoch.
It will be used to determine EpochId for an oprhaned state witness
with some height.