Split process_chunk_state_witness into two parts
Let's split this function into two parts - one part which
tries to find the previous block for the given ChunkStateWitness
and another which processes the witness when the previous block
is available.
It will make handling orphaned witnesses easier - once the block
arrives, we can just call the second function to process it.

In the future it might also make it easier to process witnesses
before the previous block is saved to the database, which could
reduce latency.