Better design for OrphanStateWitnessPool
The previous design of OrphanStateWitnessPool was overcompilcated,
let's change it to a simpler one. In the previous one I though that
I have to keep a separate small cache for every chunk producer to
prevent a single chunk producer from spamming the cache with fake
messages, but it can be solved in a better way. A simpler solution is
to keep only witnesses which are close to the current chain head.
We know that the witness is really produced by the correct chunk producer,
so the chunk producer can only take up space at the heights that they own,
without affecting heights owned by other chunk producers. Let's go with
the simpler, but still secure, solution.