Limit size of orphan witnesses in the orphan pool
Let's limit the size of orphan witnesses which are
kept in OrphanChunkStateWitnessPool to 16MB to limit
the maximum amount of memory that this pool can consume.

During StatelessNet loadtests by @staffik, the maximum
observed ChunkStateWitness size was 8MB, so 16MB should
be a safe limit. Witnesses above this size won't be saved.
This shouldn't be a problem, if a validator misses one orphaned
ChunkStateWitness they just won't vote on it, which is a normal
occurence in the protocol.

Only size of orphaned witnesses is limited. Witnesses which are
processed immediately can still be aribtrarily large. It might
not be safe to limit non-orphan witness, see the discussion in
https://github.com/near/nearcore/pull/10615